Added file
In this project, we aim to develop a smart homes temperature forecasting system. The goal is to predict the indoor temperature of smart homes based on various environmental variables and time-related features. This predictive model can be useful for homeowners to optimize their indoor climate control, enhance energy efficiency, and create a comfortable living environment.